+++ src/Devices/Ethernet_TC.h
@@ -16,7 +16,7 @@ public:
   byte *getMac() {
     return mac;
   }
-  int getNumAttemptedDHCPReleases() {
+  uint16_t getNumAttemptedDHCPReleases() {
     return numAttemptedDHCPReleases;
   };
   void renewDHCPLease();
@@ -33,9 +33,9 @@ private:
   IPAddress defaultIP = IPAddress(192, 168, 1, 2);
   IPAddress time_serverIP;
   IPAddress IP;
-  unsigned long previous_lease = 0;
-  const unsigned long LEASE_INTERVAL = 345600000;  // 4 days in milliseconds
+  uint32_t previous_lease = 0;
+  const uint32_t LEASE_INTERVAL = 345600000;  // 4 days in milliseconds
 
   // testing
-  int numAttemptedDHCPReleases = 0;
+  uint16_t numAttemptedDHCPReleases = 0;
 };
